agent_from_scratch/requirements.txt
requests 
termcolor 
python-dotenv
